microservices/img-bar/requirements.txt
Flask
requests
redis
psycopg2-binary